# remove inlined functions for faster compilation