after inlining remove functions that is not called any more
reduces compile times due to the removed functions is not further optimized.

part of solution for https://github.com/rust-lang/rust/issues/44655