@alexcrichton 2017-09-17
Wow those are some awesome wins you're seeing! Would you be willing to try to send this upstream to LLVM itself as well? Unfortunately I'm less familiar with their contribution process, but it'd be great to ensure that their experiences eyes also took a look at this!