@dcci 2017-09-18
LLVM dev here [with a hidden passion for rust] :)
I see the rationale behind your commit. I think it makes sense, and you're probably seeing a big benefit because you have a bunch of small functions (even better if marked as static) with a single caller.
If you look at the inliner code in LLVM you'll see it gives a ridicolous bonus (grep for `LastCallToStaticBonus`, or something along these lines).

I wonder whether it's better running GDCE after the inliner or having the inliner itself removing dead functions after they're inlined (admittedly, I haven't looked at whether the latter is feasible).

FWIW, what's your benchmark? Can you try with big cases (like, building rust, building clang itself, maybe with LTO [this last one generaly should leave a bunch of unused function around])