@alexcrichton 2017-09-18
@dcci hello again! FWIW on https://github.com/rust-lang/rust/issues/44655 it mentions a 23% improvement for the racer project which I think is [relatively large](https://github.com/racer-rust/racer), and it also mentions a [15% improvement for rustc itself](https://github.com/rust-lang/rust/issues/44655#issuecomment-330053207) which I know for a fact is quite large! I haven't tried this out myself yet, though, to verify.