@dcci 2017-09-18
OK, thanks for your datapoints.
FWIW, GDCE is `O(n)`, so I  don't expect that taking a lot of time, but still, we should make a more careful analysis before changing the standard `-O2` pipeline. If you're willing to run some tests on rust, I'm going to run something equivalent on C++ (as I mentioned earlier, I'm particularly interested in LTO).
I'm a little swamped at the moment so that my take a week or two before we get the patch in, how does that sound like as timeline?

Thanks!