@alexcrichton 2017-09-18
Sure yeah! @andjo403 if you want to poke around as well I can try to reproduce the rustc build times and also test some projects like Cargo and some simple-ish LTO cases.