@andjo403 2017-09-18
@alexcrichton sounds good. will try some more projects was thinking of running the https://github.com/rust-lang-nursery/rustc-benchmarks compare.py but will have to install a virtualbox or something due to it is using make and I'm on windows.