@alexcrichton 2017-09-18
Some stats:

These numbers are for whole compilations, basically how long it took `cargo build --release` to finish, which means it's timing a whole lot more than just LLVM:

* rustc itself 505s -> 423s
* cargo 215s -> 163s
* small async I/O server 57s -> 48s

Some other numbers

* Compiling Cargo with LTO, just the final rustc invocation 106s -> 95s


Definitely seems like some possible big improvements to be had! Apparently not helping LTO a whole lot though :(