@andjo403 2017-09-18
@dcci and @alexcrichton maybe I do not understand how LTO works but I do not see how functions can be removed from an other crate. the linkage must be not private for LTO to do something and it feels like GDCE only can remove private linkage functions.