@alexcrichton 2017-09-18
@andjo403 oh I think it's just with LTO that typically LLVM works with much larger LLVM modules (lots more code in memory all at once). In theory there's also tons of inlining and in theory greater wins to be had, but the one test I had didn't look too promising unfortunately :(