@dcci 2017-09-18
@alexcrichton Alex, I'm a little surprised of these numbers as, FWIW, LTO should be the kinda best case but not always really.
Please note that if you do LTO, the inliner runs twice. 1st time, during the per-TU pipeline (i.e. -O2 and -O3) and another time on the whole module.
I'm not sure how rust actually sets up the pipeline, but if it gets the stock from LLVM, you might consider adding the pass also to `populateLTOPassManager` (or something similar, I don't remember the name off the top of my head), otherwise, GDCE won't run after the inliner at link time.