@alexcrichton 2017-09-18
Aha indeed! Looks like [there's a spot](https://github.com/andjo403/llvm-1/blob/42e941e24e38c8225bad59632ce56d13949c7f38/lib/Transforms/IPO/PassManagerBuilder.cpp#L735-L740) this PR as-is doesn't handle. I've updated that and will report back the LTO build for Cargo.