@dcci 2017-09-18
Yes, that's exactly it. You were only getting the benefit from the per-TU pipeline.