@andjo403 2017-09-18
@alexcrichton there is already one a some lines below https://github.com/andjo403/llvm-1/blob/42e941e24e38c8225bad59632ce56d13949c7f38/lib/Transforms/IPO/PassManagerBuilder.cpp#L747