@dcci 2017-09-21
@alexcrichton @andjo403 
So, I've been looking at this more closely. I'm not able to find any obvious regression, so I think I'm going for a slightly modified version of this patch for review (probably, we should also run another pass, GlobalOpt in tandem with GlobalDCE). Once that is done, I guess you guys can probably backport this thingy back? Of course, I'll give the correct attribution for the idea =)