@alexcrichton 2017-09-21
@dcci sounds great, and thanks again for helping out with this! If you've got a patch I can run some tests for it for sure. And yeah we'll take care of backporting for our own "fork", no need to worry about that!