@dcci 2017-10-02
FWIW, I'm working on this, but the patch exposes a series of misoptimizations in LLVM that need to be fixed. I'm working on these now and I hope to enable the patch once they're all addressed.