@andjo403 2017-10-02
have update this PR with the patch from @dcci so it shall be easier to test for others
I have started the compilation not sure if I will wait for the results tonight