@andjo403 2017-10-02
@dcci I have only tested to compile race once with standard rustc and once with the patch and I get 
LLVM Pass execution timing report for rust lib compilation 
standard rustc  Total Execution Time: 64.7500 seconds (68.0480 wall clock)
patch Total Execution Time: 35.9531 seconds (37.2969 wall clock)

and for all crates in racer the total time go from 405.71 secs to 315.5 secs
will run some more tests tomorrow

@alexcrichton can you also run some tests?